+++ watchtower-app/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-08-11
+### Added
+- Hub Info tile: add platform update available and alerts notifications - `@amithalp`
+
+### Fixed
+- Fix Iframe tile margins - `@amithalp`
+
 ## [1.2.1] - 2024-08-11
 ### Fixed
 - Fix `crypto.randomUUID()` issue